Script Updated! new
I've made some changes to the script. It works, but it's still under work.
It bets on all questions.
It also opens crates and claims coins if on inventory page. (off by default. check the end of the script and uncomment the lines to turn it on)

Had to repost due to wrong file uploaded.